Added an AllowUninstall property to the plugins, this determines if they show up in the uninstall list. Made it so -1 id plugins can only be loaded in internal builds. Hardcoded plugin 0 and 29 0 is the core plugin and 29 is the intro tutorial.
git-svn-id: https://192.168.3.139:8443/svn/Medical/trunk@3921 956a8c8d-885e-cb46-a0b9-a2f355b4c364

+++ AnomalousMedical/GUI/DownloadManagerGUI/DownloadManagerGUI.cs
@@ -95,10 +95,7 @@ namespace Medical.GUI
                 {
                     foreach (AtlasPlugin plugin in pluginManager.LoadedPlugins)
                     {
-                        if (plugin.PluginId != -1)
-                        {
-                            installedPluginsList.Add(plugin);
-                        }
+                        installedPluginsList.Add(plugin);
                     }
                 }
                 else
@@ -107,7 +104,7 @@ namespace Medical.GUI
 
                     foreach (AtlasPlugin plugin in pluginManager.LoadedPlugins)
                     {
-                        if (plugin.PluginName != "Core")
+                        if (plugin.AllowUninstall)
                         {
                             addInfoToButtonGrid(new UninstallInfo(plugin), false);
                             installedPluginsList.Add(plugin);
@@ -115,8 +112,11 @@ namespace Medical.GUI
                     }
                     foreach (AtlasPlugin plugin in pluginManager.UnlicensedPlugins)
                     {
-                        plugin.loadGUIResources();
-                        addInfoToButtonGrid(new UninstallInfo(plugin, ServerDownloadStatus.Unlicensed), false);
+                        if (plugin.AllowUninstall)
+                        {
+                            plugin.loadGUIResources();
+                            addInfoToButtonGrid(new UninstallInfo(plugin, ServerDownloadStatus.Unlicensed), false);
+                        }
                     }
                     addedInstalledPlugins = true;
